Several changes.
F2FS now supported (I'm unsure if non-userdata partitions use F2FS, though).

The recovery now has a command to print out the partitions on the device and note which ones have filesystems, and the verifier will command it to do so if no partitions have been specified.

Symlink destinations are now considered in the verification (but at the cost of slower speed).